+++ src/main/java/com/project/open_stall/repo/ProductRepo.java
@@ -4,10 +4,11 @@ import com.project.open_stall.model.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Repository
 public interface ProductRepo extends JpaRepository<Product, Long> {
 
-    public List<Product> findByNameAndModelAndSalePriceGreaterThan(String name, String model, int salePrice);
+    public List<Product> findByNameAndModelAndSalePriceGreaterThan(String name, String model, BigDecimal salePrice);
 }
